fix:transaction-applied-v3

See merge request s13-fintech-finance-sub1/S13P21A407!177

+++ pocketc_backend/src/main/java/com/ssafy/pocketc_backend/domain/mission/dto/response/MissionDto.java
@@ -5,9 +5,9 @@ public record MissionDto(
         Integer type,
         String mission,
         String status,
-        Integer progress
+        Long progress
 ) {
-    public static MissionDto of(Integer subId, Integer type, String mission, String status, Integer progress) {
+    public static MissionDto of(Integer subId, Integer type, String mission, String status, Long progress) {
         return new MissionDto(subId, type, mission, status, progress);
     }
 }
